feat: Upgraded dependencies and use DHT to discover on same topic
This implementation will provide a hash topic over DHT while also getting providers of the same hash at a set interval and connect to them if we are not already connected.

Other small changes are a `PeerInfo` struct to track recipients peer id, addresses and public keys (encoded by crypto-seal).